Move Scheduler test to appropriate place
